Stand in (the wrong places) images
Spent waaaaaay too long trying to find a not-if-else-based way of converting my strings to variable names for sprites and I never got it to work so now I'm sleep deprived and lacking story for very little reason. But images technically exist. Hopefully I'll wake up to improve them on stream starting in 7 hours. o^o